[AI-20] move math matrix functions to Math.Matrix
